"not tired" text now appears as initially intended
pressed variable added to button classes to store whether or not they have been pressed. without it, the notTiredTimer variable would cause the condition responsible for displaying the not tired text to be true since the timer is less than 60 frames at the beginning of the game, until it is then changed to frameCount a frame later.