tested - bug fixed (read description)
the pet would not be able to sleep when fatigue was at 10% even though the "i'm not tired" message only displays when the pet is UNDER 10% fatigue (meaning 9 or less). i fixed it by displaying the bar value percentages as floored floats, but keeping the actual values the same, so how much the pet was actually fatigued would match up more accurately with the bar.