weight affects hunger
the more the pet eats, the more weight it puts on. It then becomes hungrier and needs to eat more to be satisfied, which causes it to gain more weight, and so on. This feedback loop is intentional, as it means the player has to be cautious as to not overfeed the pet, and must exercise - which teaches them that exercise is important